Opinionate defaults for CI environments
While this deviates from k3s defaults, I think they make sense for k3s
within CI environments like this will be. But, not for sure. These
options are the most lightweight options though, so I think it can be
motivated to let them be lightweight and opt in towards more
functionality as well.

Note that I just made helm-version and k3s-version blanks strings, this
is because if they are blank strings, then you get the latest version as
the environment variables set with these inputs will be ignored then.